requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
tensorflow>=2.16.1
keras>=3.0.0
scikit-learn>=1.4.0
pandas>=2.1.4
numpy>=1.26.3
Pillow>=10.2.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.25.0
psutil>=5.9.0
